Restructure entire project
- Made the crate root be the sudoku `Solver` trait along with the `Board` abstraction.
- Moved all the existing code into `solvers/psr/`. Also added `solvers/btr/` for my backtracking solution.
- All of a `Solver`s core definitions and implementations lie in their `core` submodule.
