--- src/filters/mod.rs
@@ -1,1 +1,0 @@
-// Filters are methods that remove candidates from the board
