--- src/finders/mod.rs
@@ -1,7 +1,0 @@
-//! Finders are methods that identify when a cell is known within the board.
-
-#[doc(hidden)]
-pub mod single_candidate;
-
-#[doc(hidden)]
-pub mod unique_candidate;
