+++ src/solvers/btr/mod.rs
@@ -1,0 +1,1 @@
+//! Backtracking (BTR) Solvers
