+++ src/solvers/psr/core/filters/mod.rs
@@ -1,0 +1,1 @@
+//! Methods that remove candidates from the board
