+++ src/solvers/psr/core/finders/mod.rs
@@ -1,0 +1,8 @@
+//! (methods here are taken to their underlying type)
+//! manual linking is required.
+
+#[doc(hidden)]
+pub mod single_candidate;
+
+#[doc(hidden)]
+pub mod unique_candidate;
