+++ src/solvers/psr/core/structs/cell/fmt.rs
