+++ src/solvers/psr/core/structs/idx.rs
