+++ src/solvers/psr/core/structs/idx/fmt.rs
