+++ src/solvers/psr/core/utils/grididxs.txt
