+++ src/solvers/psr/mod.rs
@@ -1,0 +1,4 @@
+//! Phase Space Reduction (PSR) Solvers
+
+// limit core to only be visible within "psr/"
+pub(super) mod core;
