--- src/solvers/simple.rs
@@ -1,3 +1,0 @@
-use super::Solver;
-
-pub struct SimpleSolver;
