+++ src/solvers/utils/boards.rs
