+++ src/solvers/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod boards;
+
